Remove generate_quote from public API
